게임 상세페이지 ssr, 자잘한 수정
