+++ app/(base)/games/[gameId]/components/WishlistButtonClient.tsx
@@ -22,16 +22,24 @@ export default function WishlistButtonClient({
     const [loading, setLoading] = useState(false);
 
     useEffect(() => {
+        console.log("📡 WishlistButtonClient viewerId:", viewerId);
+        console.log("📡 Fetching wishlists for gameId:", gameId);
+
         const fetchWishlistStatus = async () => {
             try {
                 const res = await fetch("/api/member/wishlists");
+
                 if (!res.ok) throw new Error("위시리스트 조회 실패");
 
                 const list: WishlistGame[] = await res.json();
+                console.log("✅ 현재 위시리스트:", list);
+
                 const exists = list.some((game) => game.id === gameId);
+                console.log("🧩 이 게임은 위시리스트에 있음?", exists);
+
                 setIsWished(exists);
             } catch (err) {
-                console.error("위시리스트 상태 확인 실패", err);
+                console.error("❌ 위시리스트 상태 확인 실패", err);
             }
         };
 
@@ -58,7 +66,10 @@ export default function WishlistButtonClient({
         }
     };
 
-    if (isWished === null) return null;
+    if (isWished === null) {
+        console.log("⏳ 아직 위시리스트 상태 로딩 중...");
+        return null;
+    }
 
     return (
         <Button
