+++ app/(base)/games/components/SearchBar.tsx
@@ -16,7 +16,7 @@ export default function SearchBar({ keyword, setKeyword }: SearchBarProps) {
         <div className="flex items-center gap-1">
             <div
                 className={cn(
-                    "relative w-[250px] h-[32px] rounded-[4px] border transition overflow-hidden",
+                    "relative w-[707px] h-[50px] rounded-[4px] border transition overflow-hidden",
                     isFocused ? "border-primary-purple-200" : "border-line-200"
                 )}
             >
@@ -24,8 +24,8 @@ export default function SearchBar({ keyword, setKeyword }: SearchBarProps) {
                     <Image
                         src="/icons/search.svg"
                         alt="검색"
-                        width={16}
-                        height={16}
+                        width={20}
+                        height={20}
                     />
                 </div>
 
@@ -33,7 +33,7 @@ export default function SearchBar({ keyword, setKeyword }: SearchBarProps) {
                     type="text"
                     value={keyword}
                     onChange={(e) => setKeyword(e.target.value)}
-                    placeholder="제목 혹은 개발사를 입력하세요"
+                    placeholder="영문으로 제목 혹은 개발사를 입력하세요"
                     onFocus={() => setIsFocused(true)}
                     onBlur={() => setIsFocused(false)}
                     className="w-full h-full pl-8 pr-3 text-font-100 font-small placeholder-font-200 bg-background-100 outline-none rounded-[4px]"
