+++ app/components/Pager.tsx
@@ -149,11 +149,6 @@ export default function Pager({
                                 {/* Ripple effect */}
                                 <div className="absolute inset-0 rounded-xl bg-gradient-to-r from-primary-purple-200 to-primary-blue-200 opacity-0 group-active:opacity-30 transition-opacity duration-150" />
                             </button>
-
-                            {/* Active page indicator */}
-                            {pageNumber === currentPage && (
-                                <div className="absolute -bottom-2 left-1/2 transform -translate-x-1/2 w-2 h-2 bg-gradient-to-r from-primary-purple-200 to-primary-blue-200 rounded-full animate-bounce" />
-                            )}
                         </div>
                     ))}
                 </div>
