+++ app/globals.css
@@ -36,14 +36,72 @@
 }
 
 .custom-scroll::-webkit-scrollbar {
-  width: 6px;
+    width: 6px;
 }
 
 .custom-scroll::-webkit-scrollbar-thumb {
-  background-color: rgba(100, 100, 100);
-  border-radius: 3px;
+    background-color: rgba(100, 100, 100);
+    border-radius: 3px;
 }
 
 .custom-scroll::-webkit-scrollbar-track {
-  background: rgb(47, 47, 47);
+    background: rgb(47, 47, 47);
+}
+
+@keyframes spinBackground {
+    from {
+        transform: rotate(0deg);
+    }
+    to {
+        transform: rotate(360deg);
+    }
+}
+
+.animated-border {
+    position: relative;
+    border-radius: 8px;
+    overflow: hidden;
+}
+
+.animated-border::before {
+    content: "";
+    position: absolute;
+    top: -3px;
+    left: -3px;
+    right: -3px;
+    bottom: -3px;
+    background: linear-gradient(90deg, #9333ea, #2563eb, #9333ea);
+    background-size: 300% 100%;
+    animation: borderFlow 4s linear infinite;
+    z-index: 0;
+}
+
+.animated-border-inner {
+    position: relative;
+    z-index: 1;
+    border-radius: 6px;
+    background-color: #191919; /* background-100 */
+    padding: 1px;
+}
+
+@keyframes glowFlow {
+    0% {
+        background-position: 0% 50%;
+    }
+    50% {
+        background-position: 100% 50%;
+    }
+    100% {
+        background-position: 0% 50%;
+    }
+}
+
+.glow-border {
+    position: relative;
+    border-radius: 10px;
+    padding: 3px;
+    background: linear-gradient(90deg, #9333ea, #2563eb, #9333ea, #2563eb);
+    background-size: 200% 100%;
+    animation: glowFlow 8s ease-in-out infinite;
+    box-shadow: 0 0 12px #9333ea88, 0 0 24px #2563eb66;
 }
