+++ lib/prisma.ts
@@ -1,0 +1,11 @@
+import { PrismaClient } from "@/prisma/generated"; // 이 경로 중요!!
+
+const globalForPrisma = globalThis as unknown as {
+    prisma: PrismaClient | undefined;
+};
+
+export const prisma = globalForPrisma.prisma ?? new PrismaClient();
+
+if (process.env.NODE_ENV !== "production") {
+    globalForPrisma.prisma = prisma;
+}
